chore(lib): move source files to `src` folder
