+++ l10n.yaml
@@ -1,3 +1,4 @@
+arb-dir: lib/src/l10n
 template-arb-file: intl_en.arb
 output-localization-file: query_wizard_localizations.dart
 output-class: QueryWizardLocalizations
