+++ test/tables_and_fields_test.dart
@@ -2,7 +2,7 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/query_wizard_localizations.dart';
 
-import 'package:query_wizard/tabs/tables_and_fields.dart';
+import 'package:query_wizard/src/tabs/tables_and_fields.dart';
 
 void main() {
   testWidgets('Tables and fields initialized', (WidgetTester tester) async {
